server/requirements.txt
Django==1.9.6
argparse==1.2.1
hashids==1.1.0
uWSGI==2.0.13.1
wsgiref==0.1.2
mysqlclient==1.3.7
django-taggit==0.20.2
